requirement.txt
streamlit
PyMuPDF
openai
faiss-cpu
sentence-transformers
numpy
